+++ app/Resources/views/commons/menuUser.html.twig
@@ -1,6 +1,6 @@
 <nav>
     <div class="nav-wrapper">
-        <a href="#" class="brand-logo center">
+        <a href="{{ path('index_web') }}" class="brand-logo center">
             <img class="imgLogo" src="{{ asset('img/logo.png') }}" alt="">
         </a>
         <ul id="nav-mobile" class="left">
@@ -33,7 +33,7 @@
         </div>
     </li>
     <li>
-        <a class="waves-effect" href="#">
+        <a class="waves-effect" href="{{ path('index_web') }}">
             <i class="small material-icons">home</i>Inicio</a>
     </li>
     <li>
