+++ src/AppBundle/Repository/Request_companyRepository.php
@@ -12,9 +12,6 @@ class Request_companyRepository extends \Doctrine\ORM\EntityRepository
 {
     public function getRequestCompanies()
     {
-        $qb = $this->getEntityManager()->createQueryBuilder()
-            ->select("t")
-            ->from("AppBundle:Request_company","t");
-        return $qb->getQuery()->getArrayResult();
+        return $this->findAll();
     }
 }
